Move terminal api tests into new file
